Struct cbox::CSemiBox

pub struct CSemiBox<'a, D: ?Sized> where D: DisposeRef + 'a {
    // some fields omitted
}

A wrapper for pointers made by C that are now partially owned in Rust.

This is necessary to allow owned and borrowed representations of C types to be represented by the same type as they are in C with little overhead

Methods

impl<'a, D: ?Sized> CSemiBox<'a, D> where D: DisposeRef + 'a

fn new(ptr: *mut D::RefTo) -> Self

Wrap the pointer in a CSemiBox

unsafe fn as_ptr(&self) -> *mut D::RefTo

Returns the internal pointer

unsafe fn unwrap(self) -> *mut D::RefTo

Returns the internal pointer